# gost/x/api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
    APIConfig:
        properties:
            accesslog:
                type: boolean
                x-go-name: AccessLog
            addr:
                type: string
                x-go-name: Addr
            auth:
                $ref: '#/definitions/AuthConfig'
            auther:
                type: string
                x-go-name: Auther
            pathPrefix:
                type: string
                x-go-name: PathPrefix
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    AdmissionConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            matchers:
                items:
                    type: string
                type: array
                x-go-name: Matchers
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            reverse:
                description: DEPRECATED by whitelist since beta.4
                type: boolean
                x-go-name: Reverse
            whitelist:
                type: boolean
                x-go-name: Whitelist
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    AuthConfig:
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    AutherConfig:
        properties:
            auths:
                items:
                    $ref: '#/definitions/AuthConfig'
                type: array
                x-go-name: Auths
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    BypassConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            matchers:
                items:
                    type: string
                type: array
                x-go-name: Matchers
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            reverse:
                description: DEPRECATED by whitelist since beta.4
                type: boolean
                x-go-name: Reverse
            whitelist:
                type: boolean
                x-go-name: Whitelist
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    ChainConfig:
        properties:
            hops:
                description: |-
                    REMOVED since beta.6
                    Selector *SelectorConfig `yaml:",omitempty" json:"selector,omitempty"`
                items:
                    $ref: '#/definitions/HopConfig'
                type: array
                x-go-name: Hops
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    Config:
        properties:
            admissions:
                items:
                    $ref: '#/definitions/AdmissionConfig'
                type: array
                x-go-name: Admissions
            api:
                $ref: '#/definitions/APIConfig'
            authers:
                items:
                    $ref: '#/definitions/AutherConfig'
                type: array
                x-go-name: Authers
            bypasses:
                items:
                    $ref: '#/definitions/BypassConfig'
                type: array
                x-go-name: Bypasses
            chains:
                items:
                    $ref: '#/definitions/ChainConfig'
                type: array
                x-go-name: Chains
            climiters:
                items:
                    $ref: '#/definitions/LimiterConfig'
                type: array
                x-go-name: CLimiters
            hops:
                items:
                    $ref: '#/definitions/HopConfig'
                type: array
                x-go-name: Hops
            hosts:
                items:
                    $ref: '#/definitions/HostsConfig'
                type: array
                x-go-name: Hosts
            ingresses:
                items:
                    $ref: '#/definitions/IngressConfig'
                type: array
                x-go-name: Ingresses
            limiters:
                items:
                    $ref: '#/definitions/LimiterConfig'
                type: array
                x-go-name: Limiters
            log:
                $ref: '#/definitions/LogConfig'
            metrics:
                $ref: '#/definitions/MetricsConfig'
            profiling:
                $ref: '#/definitions/ProfilingConfig'
            recorders:
                items:
                    $ref: '#/definitions/RecorderConfig'
                type: array
                x-go-name: Recorders
            resolvers:
                items:
                    $ref: '#/definitions/ResolverConfig'
                type: array
                x-go-name: Resolvers
            rlimiters:
                items:
                    $ref: '#/definitions/LimiterConfig'
                type: array
                x-go-name: RLimiters
            services:
                items:
                    $ref: '#/definitions/ServiceConfig'
                type: array
                x-go-name: Services
            tls:
                $ref: '#/definitions/TLSConfig'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    HopConfig:
        properties:
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            hosts:
                type: string
                x-go-name: Hosts
            interface:
                type: string
                x-go-name: Interface
            name:
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/NodeConfig'
                type: array
                x-go-name: Nodes
            resolver:
                type: string
                x-go-name: Resolver
            selector:
                $ref: '#/definitions/SelectorConfig'
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    HostsConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            mappings:
                items:
                    $ref: '#/definitions/HostMappingConfig'
                type: array
                x-go-name: Mappings
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    IngressConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            rules:
                items:
                    $ref: '#/definitions/IngressRuleConfig'
                type: array
                x-go-name: Rules
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    LimiterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            limits:
                items:
                    type: string
                type: array
                x-go-name: Limits
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    LogConfig:
        properties:
            format:
                type: string
                x-go-name: Format
            level:
                type: string
                x-go-name: Level
            output:
                type: string
                x-go-name: Output
            rotation:
                $ref: '#/definitions/LogRotationConfig'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    MetricsConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    ProfilingConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    RecorderConfig:
        properties:
            file:
                $ref: '#/definitions/FileRecorder'
            name:
                type: string
                x-go-name: Name
            redis:
                $ref: '#/definitions/RedisRecorder'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
    ResolverConfig:
        properties:
            name:
                type: string
                x-go-name: Name
            nameservers:
                items:
                    $ref: '#/definitions/NameserverConfig'
                type: array
                x-go-name: Nameservers
        type: object
        x-go-package: proxy_forwarder/gost/x/config
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: proxy_forwarder/gost/x/api
...
    ServiceConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            admission:
                type: string
                x-go-name: Admission
            admissions:
                items:
                    type: string
                type: array
                x-go-name: Admissions
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            climiter:
                type: string
                x-go-name: CLimiter
            forwarder:
                $ref: '#/definitions/ForwarderConfig'
            handler:
                $ref: '#/definitions/HandlerConfig'
            hosts:
                type: string
                x-go-name: Hosts
            interface:
                description: DEPRECATED by metadata.interface since beta.5
                type: string
                x-go-name: Interface
            limiter:
                type: string
                x-go-name: Limiter
            listener:
                $ref: '#/definitions/ListenerConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            recorders:
                items:
                    $ref: '#/definitions/RecorderObject'
                type: array
                x-go-name: Recorders
            resolver:
                type: string
                x-go-name: Resolver
            rlimiter:
                type: string
                x-go-name: RLimiter
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: proxy_forwarder/gost/x/config
...
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
    /config:
        get:
            operationId: getConfigRequest
            parameters:
                - description: output format, one of yaml|json, default is json.
                  in: query
                  name: format
                  type: string
                  x-go-name: Format
            responses:
                "200":
                    $ref: '#/responses/getConfigResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Get current config.
            tags:
                - Config
        post:
            operationId: saveConfigRequest
            parameters:
                - description: output format, one of yaml|json, default is yaml.
                  in: query
                  name: format
                  type: string
                  x-go-name: Format
            responses:
                "200":
                    $ref: '#/responses/saveConfigResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Save current config to file (gost.yaml or gost.json).
            tags:
                - Config
    /config/admissions:
        post:
            operationId: createAdmissionRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/AdmissionConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createAdmissionResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new admission, the name of admission must be unique in admission list.
            tags:
                - Admission
    /config/admissions/{admission}:
        delete:
            operationId: deleteAdmissionRequest
            parameters:
                - in: path
                  name: admission
                  required: true
                  type: string
                  x-go-name: Admission
            responses:
                "200":
                    $ref: '#/responses/deleteAdmissionResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete admission by name.
            tags:
                - Admission
        put:
            operationId: updateAdmissionRequest
            parameters:
                - in: path
                  name: admission
                  required: true
                  type: string
                  x-go-name: Admission
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/AdmissionConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateAdmissionResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update admission by name, the admission must already exist.
            tags:
                - Admission
    /config/authers:
        post:
            operationId: createAutherRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/AutherConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createAutherResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new auther, the name of the auther must be unique in auther list.
            tags:
                - Auther
    /config/authers/{auther}:
        delete:
            operationId: deleteAutherRequest
            parameters:
                - in: path
                  name: auther
                  required: true
                  type: string
                  x-go-name: Auther
            responses:
                "200":
                    $ref: '#/responses/deleteAutherResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete auther by name.
            tags:
                - Auther
        put:
            operationId: updateAutherRequest
            parameters:
                - in: path
                  name: auther
                  required: true
                  type: string
                  x-go-name: Auther
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/AutherConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateAutherResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update auther by name, the auther must already exist.
            tags:
                - Auther
    /config/bypasses:
        post:
            operationId: createBypassRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/BypassConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createBypassResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new bypass, the name of bypass must be unique in bypass list.
            tags:
                - Bypass
    /config/bypasses/{bypass}:
        delete:
            operationId: deleteBypassRequest
            parameters:
                - in: path
                  name: bypass
                  required: true
                  type: string
                  x-go-name: Bypass
            responses:
                "200":
                    $ref: '#/responses/deleteBypassResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete bypass by name.
            tags:
                - Bypass
        put:
            operationId: updateBypassRequest
            parameters:
                - in: path
                  name: bypass
                  required: true
                  type: string
                  x-go-name: Bypass
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/BypassConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateBypassResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update bypass by name, the bypass must already exist.
            tags:
                - Bypass
...
produces:
    - application/json
responses:
    createAdmissionResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createAutherResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createBypassResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteAdmissionResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteAutherResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteBypassResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    getConfigResponse:
        description: successful operation.
        headers:
            Config: {}
        schema:
            $ref: '#/definitions/Config'
    saveConfigResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateAdmissionResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateAutherResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateBypassResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"